update pantry feature

+++ lib/features/pantry/models/pantry_item_model.dart
@@ -1,15 +1,15 @@
 class PantryItemModel {
-	String name;
-	int quantity;
-	String unit;
-	DateTime purchaseDate;
-	DateTime expiryDate;
+  String name;
+  double quantity;
+  String unit;
+  DateTime purchaseDate;
+  DateTime expiryDate;
 
-	PantryItemModel({
-		required this.name,
-		required this.quantity,
-		required this.unit,
-		required this.purchaseDate,
-		required this.expiryDate,
-	});
+  PantryItemModel({
+    required this.name,
+    required this.quantity,
+    required this.unit,
+    required this.purchaseDate,
+    required this.expiryDate,
+  });
 }
